requirements.txt
yt-dlp
openai-whisper
google-generativeai
python-dotenv
# torch, torchvision, torchaudio sould be isntalled with openai-whisper's dependency
# if you have NVIDIA GPU and look for CUDA GPU support, please look for certain command: https://pytorch.org/get-started/locally/
# for me, on windows with RTX 2060, the command would be: pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128